+++ portal/test_settings.py
@@ -4,8 +4,6 @@ from settings import *
 from local_settings import *
 
 
-USE_TZ = True
-
 # This can be useful to debug segmentation fault errors
 # faulthandler.enable()
 
@@ -13,4 +11,3 @@ try:
     from local_test_settings import *
 except ImportError:
     pass
-